refactors the sqlexecute
